consolidate barChart constants
