--- packages/sanddance/src/specs/barchartH/constants.ts
@@ -1,16 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license.
-
-export const BarChartScaleNames = {
-    bucketScale: 'bucketScale',
-    levelScale: 'levelScale',
-    compartmentScale: 'compartmentScale'
-};
-
-export const BarChartSignalNames = {
-    aspectRatioSignal: 'aspectRatioSignal',
-    compartmentsPerLevelSignal: 'compartmentsPerLevelSignal',
-    compartmentHeightSignal: 'compartmentHeightSignal',
-    levelExtentSignal: 'levelExtentSignal',
-    quantitativeBinSignal: 'quantitativeBinSignal'
-};
