+++ packages/sanddance/src/specs/barchartV/transform.quantitative.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-import { BarChartSignalNames } from './constants';
+import { BarChartSignalNames } from '../constants';
 import { Column, SpecContext } from '../types';
 import { FieldNames, SignalNames } from '../constants';
 import { StackTransform, Transforms } from 'vega-typings';
